refactor(cli): organize admin commands
