+++ app/src/cli/commands/admin/mod.rs
@@ -1,7 +1,10 @@
+mod args;
 mod handler;
 
+pub use args::AdminCommand;
+pub(crate) use args::HELP;
+pub(super) use handler::execute;
 pub use handler::{
   complete_factory_reset, factory_reset_archive_path, factory_reset_confirmation_matches,
   factory_reset_quarantine_path, validate_factory_reset_target,
 };
-pub(super) use handler::execute;
