+++ app/src/constants/help.rs
@@ -26,20 +26,6 @@ Common server options:
 Run 'dopbase help <command>' for details on any command.
 ";
 
-pub(crate) const ADMIN_HELP: &str = "\
-Examples:
-  dopbase admin reset-password admin@example.com
-  dopbase admin factory-reset
-";
-pub(crate) const ADMIN_RESET_PASSWORD_HELP: &str =
-  "Examples:\n  dopbase admin reset-password admin@example.com\n";
-pub(crate) const ADMIN_FACTORY_RESET_HELP: &str = "\
-Examples:
-  dopbase admin factory-reset
-  dopbase admin factory-reset --no-backup
-  dopbase --data-dir /srv/dopbase admin factory-reset
-";
-
 pub(crate) const UPDATE_HELP: &str = "Examples:\n  dopbase update\n";
 
 pub(crate) const ENVIRONMENT_ARG_HELP: &str = "Existing environment reference: an environment ID or \
